+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## Changelog
 
 ### Version 1.0.2
+- **Feature**: Added new entity data storage for request and response queries.
 - **Compatibility**: Magento 2.4.3 compatibility and general code improvement.
 - **Compatibility**: Compatibility with removed \Laminas\Log\Loger package.